+++ src/commands/tokenizedBASICFile.ts
@@ -1,0 +1,4 @@
+export interface TokenizedBASICFile {
+	outputFile: string;
+	outputDir: string;
+  };