+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to the "c64basicv2" extension will be documented in this file.
 
+## [0.4.0] - 10/04/2022
+
+- [Commands] Added "Convert" and "Convert and Run" commands
+- [Keyboard Shortcuts] Added keyboard shortcuts for commands
+
 ## [0.3.5] - 05/04/2022
 
 - [Syntax highlighting] Removed word boundary controls to support lines with keywords without separators. Ex: 350 ifx=1thenprintchr$(18);